Combine ReactJSXElementValidator with main module
There are too many layers to the JSX runtime implementation. I think
basically everything should be implemented in a single file, so that's
what I'm going to do.

As a first step, this deletes ReactJSXElementValidator and moves all
the code into ReactJSXElement. I can already see how this will help
us remove more indirections in the future.

Next I'm going to do start moving the `createElement` runtime into
this module as well, since there's a lot of duplicated code.
